refactor(browser): extract tab favicon renderer
