+++ src/components/browser-pane.tsx
@@ -15,58 +15,7 @@ import {
   type BrowserNavigationRequest,
   type ExpectedBrowserNavigation,
 } from "@/lib/browser-navigation-queue";
-
-// ── Favicon helpers (mirrors open-sesame FaviconService pattern) ──────────────
-// Primary: Google S2 API (works from renderer, no CORS)
-// Fallback: colored initial chip (mirrors ColoredInitialAvatar from open-sesame)
-
-function faviconUrl(url: string): string {
-  try {
-    const host = new URL(url).hostname;
-    return `https://www.google.com/s2/favicons?domain=${host}&sz=32`;
-  } catch {
-    return "";
-  }
-}
-
-const INITIAL_COLORS = [
-  "#5b5bd6", "#7c3aed", "#db2777", "#ea580c",
-  "#16a34a", "#0891b2", "#4f46e5", "#0d9488",
-];
-
-function initialColor(name: string): string {
-  let hash = 0;
-  for (let i = 0; i < name.length; i++) hash = (hash * 31 + name.charCodeAt(i)) | 0;
-  return INITIAL_COLORS[Math.abs(hash) % INITIAL_COLORS.length];
-}
-
-function TabFavicon({ url, title, size = 20 }: { url: string; title: string; size?: number }) {
-  const [failed, setFailed] = useState(false);
-  const src = faviconUrl(url);
-  const initial = (title || url).trim().slice(0, 1).toUpperCase() || "?";
-  const color = initialColor(title || url);
-
-  if (!src || failed) {
-    return (
-      <span
-        className="flex shrink-0 items-center justify-center rounded-[var(--radius-control)] text-[length:var(--text-2xs)] font-semibold text-white"
-        style={{ width: size, height: size, background: color }}
-      >
-        {initial}
-      </span>
-    );
-  }
-  return (
-    <img
-      src={src}
-      alt=""
-      width={size}
-      height={size}
-      className="rounded-[var(--radius-control)] object-contain [image-rendering:auto]!"
-      onError={() => setFailed(true)}
-    />
-  );
-}
+import { TabFavicon } from "./browser-tab-favicon";
 
 // Browser pane — uses Tauri's child WebviewBuilder under the hood. A real
 // Chromium webview is overlaid on top of the placeholder <div> below; we
